Optimizations

+++ Syntherrupter_Tiva/EEPROMLayouts.h
@@ -187,7 +187,7 @@ public:
         constexpr uint8_t LAST = ENV_DATA_POINTS - 1;
 
         // Program 0 and all unspecified programs: No envelope (constant 100% volume while on, no rise/fall times)
-        for (uint8_t step = 0; step <= LAST; step++)
+        for (uint32_t step = 0; step <= LAST; step++)
         {
             layout.envelopes[0][step] = {.amplitude = 1.0f, .durationUS = 1.0f, .ntau = 0.1f, .nextStep = (uint8_t) (step + 1)};
         }
